moved everything from Chat.js to ChannelList.js
