+++ docs/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"><meta http-equiv="X-UA-Compatible" content="ie=edge"><title>React Chat - Docs</title><link rel="icon" type="image/x-icon" href="https://getstream.imgix.net/images/favicons/favicon-96x96.png"></head><body><div id="rsg-root"></div><script src="build/bundle.eeb02426.js"></script></body></html>
+<!DOCTYPE html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"><meta http-equiv="X-UA-Compatible" content="ie=edge"><title>React Chat - Docs</title><link rel="icon" type="image/x-icon" href="https://getstream.imgix.net/images/favicons/favicon-96x96.png"></head><body><div id="rsg-root"></div><script src="build/bundle.d752e3c3.js"></script></body></html>
